Simplify stdout, stderr in kopia runner
Previous implementation was intended to log stdout as it arrived, but
since the --no-progress flag was set and the robustness job
is running in shippable anyway, there's no need for live
output from the kopia run command.

Instead simplify the output capture more like the fixed setup in
cli_test_env.go: set the command Stderr to a bytes.Buffer and
run it by c.Output to capture stdout.