Return actual blob state at the specified PiT.
Return the blob version at time=pit, regardless of whether the blob had been
previously deleted. This removes the kopia-specific assumptions that a blob is
written only once.

Prior to this change, if a blob Y had the following versions:

Y = [ t4=v4, t3=v3, t2=d, t1=v1, t0=v0 ]

where 'd' is a deletion marker. Then, when the point-in-time t was

- t0 <= t < t1 => v0 was returned
- t1 <= t < t2 => v1 was returned
- t2 <= t => no blob was returned (Not found)

For example, if t > t4, blob not found was returned.

With this change, when the point-in-time t is:

- t0 <= t < t1 => v0 is returned
- t1 <= t < t2 => v1 is returned
- t2 <= t < t3 => no blob is returned (Not found)
- t3 <= t < t4 => v3 is returned
- t4 <= t => v4 is returned

Motivation: this behavior is less surprising as it reflects the actual state of
the store at time `t`. Also, this removes a form of interference with various
tests.